service-auth/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
python-multipart
